requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
withings-api==2.3.0
schedule==1.2.0
python-dateutil==2.8.2
pandas==2.1.0
numpy==1.24.3
redis==5.0.0
plotly==5.16.1
dash==2.13.0
dash-bootstrap-components==1.5.0
scikit-learn==1.3.0
python-telegram-bot==20.5 